requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
plotly>=5.15.0
scikit-learn>=1.3.0

# Data providers
yfinance>=0.2.20
requests>=2.31.0

# Environment and configuration
python-dotenv>=1.0.0

# Optional: For enhanced data sources
alpha-vantage>=2.3.1  # Uncomment if using Alpha Vantage API
